+++ compose/ui/ui/src/webMain/kotlin/androidx/compose/ui/draganddrop/WebDragAndDropManager.kt
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2025 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package androidx.compose.ui.draganddrop
 
 import androidx.compose.ui.events.EventTargetListener
@@ -12,6 +28,7 @@ import androidx.compose.ui.platform.PlatformDragAndDropSource
 import androidx.compose.ui.scene.ComposeSceneDragAndDropNode
 import androidx.compose.ui.unit.Density
 import androidx.compose.ui.unit.LayoutDirection
+import kotlin.js.js
 import kotlin.math.roundToInt
 import kotlinx.browser.document
 import kotlinx.browser.window
@@ -163,7 +180,10 @@ internal abstract class WebDragAndDropManager(private val rootNode: Node, eventL
 }
 
 @Suppress("UNUSED_PARAMETER")
-private fun setMethodImplForUint8ClampedArray(obj: Uint8ClampedArray, index: Int, value: Int) { js("obj[index] = value;") }
+private fun setMethodImplForUint8ClampedArray(obj: Uint8ClampedArray, index: Int, value: Int) {
+    js("obj[index] = value;")
+}
+
 private operator fun Uint8ClampedArray.set(index: Int, value: Int) = setMethodImplForUint8ClampedArray(this, index, value)
 
 private fun IntArray.toUint8ClampedArray(): Uint8ClampedArray {
